+++ src/react-components/components/Icons/IconTypes.tsx
@@ -1,4 +1,4 @@
-export enum iconRotationTypes {
+export enum IconRotationTypes {
     rotate180 = "rotate-180",
     rotate270 = "rotate-270",
     rotate0 = "rotate-0",