+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 Currently, this repo is in Prerelease. When it is released, this project will adhere to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 ========
 
+## [0.4.1]
+
+-   Moves tests into their corresponding component folders
+
+## [0.4.0]
+
+### Breaking Changes
+
+-   Concatenates `BasicLink`, `UnderlineLink`, and `IconLink` into `Link` with `LinkTypes` enum
+
 ## [0.3.7]
 
 -   Fixed `dist` script so it clears the folder